Setter Fee Transaction in X-Chain

+++ app.ts
@@ -36,7 +36,7 @@ process.env.NODE_TLS_REJECT_UNAUTHORIZED = "0";
 
 /*global variables*/
 var pathGrungni = process.env.GRUNGNI_PATH;
-let privateKeys: string[] = [];
+let privateKeys: any[] = [];
 let chainId: number = 0;
 let gasPrice: number;
 let balance: string;
@@ -57,6 +57,12 @@ let txBuilderB: ITransactionBuilder;
 let networkRunner: NetworkRunner;
 
 let web3: Web3;
+let urlRpcDetails: URL;
+let protocolRPC: string;
+let chainType: string;
+let configDataFlow: DataFlow;
+
+
 const app = express();
 app.use('/public', express.static('public'))
 app.use(bodyParser.urlencoded({ extended: false }));
@@ -73,11 +79,13 @@ app.post("/start", async (req, res) => {
     let completeTestConfiguration: ConfigurationTypeForCompleteTest = req.body;
 
     //extract last route from url
-    var networkName = completeTestConfiguration.rpc.split("/").pop();
+    //var networkName = completeTestConfiguration.rpc.split("/").pop();
+    var network = completeTestConfiguration.rpc.split("/")
+    var networkName = (network[2].split("."))[0]
 
     //read json file
     var jsonData: any = JSON.parse(fs.readFileSync(pathGrungni + "/" + networkName + ".json", "utf8"));
-    var privateKeyFirstStaker = jsonData.Stakers[0].PrivateKey;    
+    var privateKeyFirstStaker = jsonData.Stakers[0].PrivateKey;
     //cast into configurationtype
     let configType: ConfigurationType = completeTestConfiguration as ConfigurationType;
     configType.private_key_with_funds = privateKeyFirstStaker;
@@ -100,8 +108,16 @@ app.post("/start", async (req, res) => {
         console.log("Running case .. " + i);
         await initNetwork(testCase, prevTestCase, networkName, configType);
         var dataFlow = await initApp(configType);
+        if (testCase.Chain == "X") {
+            let xChainAvalanche = await getXKeyChain(urlRpcDetails.hostname, parseInt(urlRpcDetails.port), protocolRPC, dataFlow.networkID, configType.private_key_with_funds, dataFlow.assetID, dataFlow.blockchainIDXChain);
+            let mainAccount = new XChainTestWallet(dataFlow.bech32_xchain_address, configType.private_key_with_funds, xChainAvalanche);
+            utils.xChainAvalanche = xChainAvalanche;
+            utils.mainAccount = mainAccount;
+            chainType = testCase.Chain;
+
+        }
         await initPrivateKeys(dataFlow, testCase);
-        await startTestsAndGatherMetrics(testCase, configType);
+        await startTestsAndGatherMetrics(testCase, configType, networkName);
     }
 
     console.log("Finished all tests");
@@ -117,119 +133,27 @@ app.post("/network-runner", async (req, res) => {
     for (let i = 0; i < testCases.length; i++) {
         let testCase = testCases[i];
         networkRunner = await getConfigTypeWithNetworkRunner(req, testCase);
+        console.log(networkRunner);
+
         let configType = networkRunner.configuration;
         let dataFlow = await initApp(networkRunner.configuration);
+        if (testCase.Chain == "X") {
+            let xChainAvalanche = await getXKeyChain(urlRpcDetails.hostname, parseInt(urlRpcDetails.port), protocolRPC, dataFlow.networkID, networkRunner.configuration.private_key_with_funds, dataFlow.assetID, dataFlow.blockchainIDXChain);
+            let mainAccount = new XChainTestWallet(dataFlow.bech32_xchain_address, networkRunner.configuration.private_key_with_funds, xChainAvalanche);
+            utils.xChainAvalanche = xChainAvalanche;
+            utils.mainAccount = mainAccount;
+            chainType = testCase.Chain;
 
-        await initPrivateKeys(dataFlow, networkRunner.testCase);
-        await startTestsAndGatherMetrics(networkRunner.testCase, configType);
+        }
+
+        await initPrivateKeys(dataFlow, testCase);
+
+        await startTestsAndGatherMetrics(testCase, configType, "");
         await networkRunner.killGnomeTerminal();
     }
 
     return res.status(200).send("Network Runner executed");
 });
-
-app.post("/network-runner/x-chain-test", async (req, res) => {
-    let testCase: TestCase = {
-        StakingAmount: req.body.StakingAmount,
-        ValidatorNodes: req.body.ValidatorNodes,
-        ApiNodes: req.body.ApiNodes,
-        Threads: req.body.Threads,
-        Loops: req.body.Loops,
-        Chain: "X",
-        TestType: req.body.TestType,
-        Position: 0
-    }
-
-    networkRunner = await getConfigTypeWithNetworkRunner(req, testCase);
-
-    web3 = new Web3(networkRunner.configuration.rpc + '/ext/bc/C/rpc');
-
-    await Utils.createUserAccount(networkRunner.configuration);
-
-    //Information Data, asset and protocol
-    let assetID = await Utils.getStakingAssetID(networkRunner.configuration);
-    let networkID = parseInt(await Utils.getNetworkID(networkRunner.configuration));
-    let url = new URL(networkRunner.configuration.rpc);
-    let port = "443";
-    let protocolRPC = url.protocol.replace(":", "");
-    if (url.port != "") {
-        port = url.port;
-    }
-
-    console.log("URL -> ", url);
-
-
-    let accountAVM = await Utils.ImportKeyAVM(networkRunner.configuration.private_key_with_funds, networkRunner.configuration);
-    let xChainAvalanche = await getXKeyChain(url.hostname, parseInt(url.port), protocolRPC, networkID, networkRunner.configuration.private_key_with_funds, assetID);
-    let accountsXChain: XChainTestWallet[] = [];
-    let promisesXChainWallet = [];
-    let principalAccount = new XChainTestWallet(accountAVM, networkRunner.configuration.private_key_with_funds, xChainAvalanche);
-
-
-
-    //Create Private Keys and Wallets
-    for (let x = 0; x < networkRunner.testCase.Threads; x++) {
-        promisesXChainWallet.push(XChainTestWallet.importKeyAndCreateWallet(web3, networkRunner, url, protocolRPC, networkID, assetID));
-    }
-    accountsXChain = await Promise.all(promisesXChainWallet);
-
-
-
-    //Funds Wallets
-    let baseAmount: number = 400000000000000;
-    let baseSend: number = 400000000;
-    let promiseFund = [];
-
-    let c = 0;
-    while (c <= networkRunner.testCase.Threads) {
-        if (c < 10) //First 10 Transactions
-        {
-            console.log("First Transactions -> ", c)
-            await Utils.sendTransactionXChain(principalAccount, accountsXChain[c], baseAmount, principalAccount.avalancheXChain);
-        }
-        else {
-            if (c % 10 == 0 && promiseFund.length > 0) {
-                console.log("Iteration --->", (c / 10));
-                await Promise.all(promiseFund);
-                promiseFund = [];
-                baseSend = baseSend - 2000000;
-
-                if (c >= networkRunner.testCase.Threads) {
-                    console.log("BREAKING");
-                    break;
-                }
-            }
-
-            baseSend = baseSend + c;
-
-            promiseFund.push(Utils.sendTransactionXChain(accountsXChain[c - 10], accountsXChain[c], baseSend, accountsXChain[c - 10].avalancheXChain));
-        }
-        c++;
-    }
-
-    console.log("Finished Do Funds");
-
-    let promiseTransactions = [];
-
-    //Repeat Cicle for start transactions
-    for (let k = 0; k < accountsXChain.length - 1; k++) {
-        let oldBalance = await accountsXChain[k].avalancheXChain.xchain.getBalance(accountsXChain[k].xChainAddress, accountsXChain[k].avalancheXChain.avaxAssetID);
-        console.log("Old Balance ->",oldBalance);
-        console.log("K Position ->",k);
-        console.log("Address From->", accountsXChain[k].xChainAddress);
-        console.log("Address To->", accountsXChain[k + 1].xChainAddress);
-        console.log(accountsXChain[k].avalancheXChain.xKeyChain.getAddressStrings());
-        await Utils.sendTransactionXChain(accountsXChain[k], accountsXChain[k + 1], 2000000, accountsXChain[k].avalancheXChain);
-    }
-
-    console.log("Starting Transactions....");
-
-    //await Promise.all(promiseTransactions);
-    await networkRunner.killGnomeTerminal();
-    return res.status(200).send("Execution Test Finished");
-
-});
-
 
 // single endpoint to test (cloud)
 app.post('/', async (req, res) => {
@@ -240,14 +164,45 @@ app.post('/', async (req, res) => {
     if (req.body.ID == privateKeys.length) {
         sendTo = privateKeys[0];
     }
+    if (chainType == "X") {
 
-    txBuilder.buildAndSendTransaction(privateKey, contractAddress, sendTo, Constants.AMOUNT_TO_TRANSFER)
-        .then(data => {
-            res.send(data);
-        }).catch(err => {
-            errorLogger.error(err);
-            res.status(500).send(err);
-        });
+        //TODO: Temporal Blockchain ID in Addresses, change to Dynamic Blockchain ID but using Avalanche Js
+        privateKey.xChainAddress = privateKey.xChainAddress.replace("X-",`${configDataFlow.blockchainIDXChain}-`);
+        sendTo.xChainAddress = sendTo.xChainAddress.replace("X-",`${configDataFlow.blockchainIDXChain}-`);
+
+        let xWallet: XChainTestWallet = privateKey;
+        let balance = await xWallet.avalancheXChain.xchain.getBalance(xWallet.xChainAddress, xWallet.avalancheXChain.avaxAssetID);
+
+        console.log("______________________________________________");
+        console.log("Private Key ->", xWallet.privateKey);
+        console.log("ID", req.body.ID);
+        console.log("Address From:", privateKey.xChainAddress);
+        console.log("Address to:", sendTo.xChainAddress);
+        console.log("Amount:", web3.utils.toWei(Constants.AMOUNT_TO_TRANSFER, 'gwei'));
+        console.log("Balance:", balance);
+
+        //Temporal Amount
+        let ammountConversion = web3.utils.toWei(Constants.AMOUNT_TO_TRANSFER, 'gwei');
+        let xChainAvalanche = await getXKeyChain(urlRpcDetails.hostname, parseInt(urlRpcDetails.port), protocolRPC, configDataFlow.networkID, privateKey.privateKey, configDataFlow.assetID,configDataFlow.blockchainIDXChain);
+        txBuilder.buildAndSendTransaction(privateKey, contractAddress, sendTo, ammountConversion, xChainAvalanche)
+            .then(data => {
+                res.send(data);
+            }).catch(err => {
+                errorLogger.error(err);
+                res.status(500).send(err);
+            });
+
+    }
+    else {
+        txBuilder.buildAndSendTransaction(privateKey, contractAddress, sendTo, Constants.AMOUNT_TO_TRANSFER)
+            .then(data => {
+                res.send(data);
+            }).catch(err => {
+                errorLogger.error(err);
+                res.status(500).send(err);
+            });
+    }
+
 });
 
 app.post('/pingpong', async (req, res) => {
@@ -294,8 +249,10 @@ async function initApp(configType: ConfigurationType): Promise<DataFlow> {
     console.log("Block number: ", blockNumber);
 
     let dataFlow = await initDataFlowAccount(configType);
-    utils = new Utils(configType, dataFlow);
     await initBuilder(configType, dataFlow);
+    utils = new Utils(configType, dataFlow);
+    utils.urlRpc = urlRpcDetails;
+    utils.protocolRPC = protocolRPC;
     return dataFlow;
 }
 
@@ -307,6 +264,9 @@ async function initDataFlowAccount(configurationtype: ConfigurationType): Promis
 
     var hexPK = Utils.translateCb58PKToHex(configurationtype.private_key_with_funds);
     let account = web3.eth.accounts.privateKeyToAccount('0x' + hexPK);
+    let assetID = await Utils.getStakingAssetID(configurationtype);
+    let networkID = parseInt(await Utils.getNetworkID(configurationtype));
+    let blockchainIDXChain = await Utils.getBlockchainID(configurationtype, "X");
     balance = await web3.eth.getBalance(account.address);
 
     console.log("Account: ", account.address);
@@ -320,9 +280,12 @@ async function initDataFlowAccount(configurationtype: ConfigurationType): Promis
         bech32_cchain_address: bech32AddressXChain.replace("X", "C"),
         hex_cchain_address: account.address,
         gasPrice: gasPrice.toString(),
-        chainId: chainId
+        chainId: chainId,
+        assetID: assetID,
+        networkID: networkID,
+        blockchainIDXChain: blockchainIDXChain
     }
-
+    configDataFlow = dataFlow;
     console.log("Data flow: ", dataFlow);
     return dataFlow;
 }
@@ -330,46 +293,60 @@ async function initDataFlowAccount(configurationtype: ConfigurationType): Promis
 
 // function to initialize the app
 async function initPrivateKeys(dataflow: DataFlow, testCase: TestCase): Promise<Boolean> {
-
-    //evaluate if file exists
-    if (fs.existsSync(Constants.PRIVATE_KEYS_FILE)) {
-        privateKeys = fs.readFileSync(Constants.PRIVATE_KEYS_FILE).toString().split("\n");
-        let account = web3.eth.accounts.privateKeyToAccount(privateKeys[0]);
-        balance = await web3.eth.getBalance(account.address);
-        if (balance == "0") {
-            //delete file privatekeys.csv
-            fs.unlinkSync(Constants.PRIVATE_KEYS_FILE);
+    if (testCase.Chain == "C") {
+        //evaluate if file exists
+        if (fs.existsSync(Constants.PRIVATE_KEYS_FILE)) {
+            privateKeys = fs.readFileSync(Constants.PRIVATE_KEYS_FILE).toString().split("\n");
+            let account = web3.eth.accounts.privateKeyToAccount(privateKeys[0]);
+            balance = await web3.eth.getBalance(account.address);
+            if (balance == "0") {
+                //delete file privatekeys.csv
+                fs.unlinkSync(Constants.PRIVATE_KEYS_FILE);
+            }
         }
-    } 
-    if (balance == "0") {
-        //import and export
-        await utils.transferFunds();
-        balance = await web3.eth.getBalance(dataflow.hex_cchain_address);
-        console.log('New balance after import/export : ', balance);
+        if (balance == "0") {
+            //import and export
+            await utils.transferFunds();
+            balance = await web3.eth.getBalance(dataflow.hex_cchain_address);
+            console.log('New balance after import/export : ', balance);
+        }
+        if (testCase.TestType == "erc20tx" || testCase.TestType == "erc1155tx") {
+            contractAddress = await txBuilder.deployContract("0x" + dataflow.hexPrivateKey, web3);
+            txBuilder.contractAddress = contractAddress;
+            utils.txBuilder = txBuilder;
+        }
+        // initialize accounts
+        console.log("Generating accounts ... ");
+        await utils.generateAndFundWallets(testCase);
+        // read file private keys using fs
+        privateKeys = fs.readFileSync(Constants.PRIVATE_KEYS_FILE).toString().split("\n");
     }
-    if (testCase.TestType== "erc20tx" || testCase.TestType == "erc1155tx"){
-        contractAddress = await txBuilder.deployContract("0x"+dataflow.hexPrivateKey, web3);
-        txBuilder.contractAddress = contractAddress;
-        utils.txBuilder = txBuilder;
+    //private keys create wallets and send funds in xchain 
+    else {
+        // initialize accounts
+        console.log("Generating accounts ... ");
+        await utils.generateAndFundWallets(testCase, txBuilder);
+        privateKeys = utils.privateKeys;
     }
-    // initialize accounts
-    console.log("Generating accounts ... ");
-    await utils.generateAndFundWallets(testCase);
-    // read file private keys using fs
-    privateKeys = fs.readFileSync(Constants.PRIVATE_KEYS_FILE).toString().split("\n");
+
 
     return true;
 
 }
 
 async function initBuilder(configurationType: ConfigurationType, dataFlow: DataFlow) {
-    
+
     // initialize transaction builder
     switch (configurationType.test_type) {
         case "transfer": txBuilder = new SimpleTXBuilder(configurationType, web3, dataFlow);
             // mint
             break;
         case "erc20tx": txBuilder = new testbuilderErc20(configurationType, web3, dataFlow);
+            break;
+        case "transfer-xchain":
+            txBuilder = new xChainBuilder(configurationType, web3, dataFlow);
+            urlRpcDetails = await getURLDetails(configurationType.rpc);
+            protocolRPC = urlRpcDetails.protocol.replace(":", "");
             break;
         case "erc1155tx": txBuilder = new ERC1155TXBuilder(configurationType, web3, dataFlow);
             break;
@@ -378,6 +355,14 @@ async function initBuilder(configurationType: ConfigurationType, dataFlow: DataF
     }
 }
 
+async function getURLDetails(rpc: string) {
+    let url = new URL(rpc);
+    let port = "443";
+    if (url.port != "") {
+        port = url.port;
+    }
+    return url;
+}
 
 async function initNetwork(testCase: TestCase,
     prevTestCase: TestCase,
